+++ CHANGELOG.md
@@ -5,11 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.20.0] - 2022-11-16
+
+### Added
+
+-   Added missing base meta fields:
+    -   Homepage
+    -   Author
+    -   Separator
+    -   404
+
 ## [4.19.0] - 2022-10-21
 
 ### Added
 
--   Added "Other" Soical Profiles after Yoast change
+-   Added "Other" Social Profiles after Yoast change
 
 ## [4.18.0] - 2022-06-10